third_party/blink/web_tests/fast/forms/range/range-keyoperation-expected.txt

Test for keyboard operations of <input type=range>

Tests for a horizontal range
Press the up arrow key:
PASS input.value is "51"
PASS changeEventCounter is lastChangeEventCounter + 1
Press the down arrow key:
PASS input.value is "50"
Press the left arrow key:
PASS input.value is "49"
Press the right arrow key:
PASS input.value is "50"
Press the PageUp key:
PASS input.value is "60"
Press the PageDown key:
PASS input.value is "50"
Press the Home key:
PASS input.value is "0"
Press the End key:
PASS input.value is "100"
Edge cases
PASS input.value is "0"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "1"
PASS input.value is "100"
PASS input.value is "99"
PASS input.value is "0"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "0"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "100"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "100"
PASS changeEventCounter is lastChangeEventCounter


Tests for a horizontal range with RTL
Press the up arrow key:
PASS input.value is "51"
Press the down arrow key:
PASS input.value is "50"
Press the left arrow key:
PASS input.value is "51"
Press the right arrow key:
PASS input.value is "50"


Tests for a vertical range
Press the up arrow key:
PASS input.value is "51"
Press the down arrow key:
PASS input.value is "50"
Press the left arrow key:
PASS input.value is "49"
Press the right arrow key:
PASS input.value is "50"
Press the PageUp key:
PASS input.value is "60"
Press the PageDown key:
PASS input.value is "50"
Press the Home key:
PASS input.value is "0"
Press the End key:
PASS input.value is "100"


Tests for a vertical range with rtl
Press the up arrow key:
PASS input.value is "51"
Press the down arrow key:
PASS input.value is "50"
Press the left arrow key:
PASS input.value is "51"
Press the right arrow key:
PASS input.value is "50"
Press the PageUp key:
PASS input.value is "60"
Press the PageDown key:
PASS input.value is "50"
Press the Home key:
PASS input.value is "0"
Press the End key:
PASS input.value is "100"


step=any cases
Press the up arrow key:
PASS input.value is "102"
PASS changeEventCounter is lastChangeEventCounter + 1
Press the down arrow key:
PASS input.value is "100"
PASS changeEventCounter is lastChangeEventCounter + 1
Press the pageup key:
PASS input.value is "120"
PASS changeEventCounter is lastChangeEventCounter + 1
Press the pagedown key:
PASS input.value is "100"
PASS changeEventCounter is lastChangeEventCounter + 1
Press the home key:
PASS input.value is "0"
PASS changeEventCounter is lastChangeEventCounter + 1
Press the end key:
PASS input.value is "200"
PASS changeEventCounter is lastChangeEventCounter + 1
Edge cases
PASS input.value is "200"
PASS input.value is "200"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "198"
PASS input.value is "200"
PASS input.value is "200"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "180"
PASS input.value is "0"
PASS input.value is "0"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "2"
PASS input.value is "0"
PASS input.value is "0"
PASS changeEventCounter is lastChangeEventCounter
PASS input.value is "20"


small range
PASS input.value is "9"
PASS input.value is "6"
PASS input.value is "9"
PASS input.value is "0"
PASS input.value is "9"


Disabled and read-only
PASS input.value is "1"
PASS input.value is "1"
PASS input.value is "1"
PASS input.value is "1"
PASS input.value is "1"
PASS input.value is "1"


PASS successfullyParsed is true

TEST COMPLETE

      
